chore: Migrate SqlLab API to API v1
